Ensure bin at application startup
