+++ action/compose_test.go
@@ -72,29 +72,6 @@ version: "3"
 
 }
 
-func TestEnsureBin(t *testing.T) {
-	tests := map[string]struct {
-		input      string
-		shouldFail bool
-	}{
-		"valid":   {input: "docker-compose", shouldFail: false},
-		"invalid": {input: "dckr-cmps", shouldFail: true},
-	}
-
-	for name, tc := range tests {
-		t.Run(name, func(t *testing.T) {
-			err := ensureBin(tc.input)
-			if tc.shouldFail {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-
-		})
-	}
-
-}
-
 func TestRun(t *testing.T) {
 	c, err := NewValidCompose()
 	assert.NoError(t, err)
