+++ cmd/batch-scheduler.go
@@ -5,10 +5,16 @@ import (
 	"log"
 	"time"
 
+	"github.com/factorysh/batch-scheduler/runners"
 	"github.com/factorysh/batch-scheduler/server"
 )
 
 func main() {
+
+	err := runners.EnsureBin("docker-compose")
+	if err != nil {
+		log.Fatal(err)
+	}
 
 	var s server.Server
 
@@ -18,7 +24,7 @@ func main() {
 	<-s.Done
 
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	err := s.API.Shutdown(ctx)
+	err = s.API.Shutdown(ctx)
 	defer func() {
 		cancel()
 	}()
